# api-docs.yaml
openapi: 3.0.0
info:
  title: IndoRates Pro API
  description: Dokumentasi API dengan CRUD Lengkap (MySQL + Prisma).
  version: 2.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []

paths:
  # --- AUTHENTICATION ---
  /api/auth/register:
    post:
      summary: Daftar Akun Baru (Create)
      security: []
      tags: [Authentication]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        200: {description: Berhasil daftar}

  /api/auth/login:
    post:
      summary: Login User
      security: []
      tags: [Authentication]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        200: {description: Mengembalikan API Key}

  # --- FEATURES ---
  /api/rates:
    get:
      summary: Baca Data Kurs (Read)
      tags: [Features]
      responses:
        200: {description: List kurs terbaru}

  /api/convert:
    post:
      summary: Kalkulator Konversi
      tags: [Features]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                currency: {type: string}
                amount: {type: number}
      responses:
        200: {description: Hasil hitungan}

  # --- ADMIN CRUD ---
  /api/admin/users:
    get:
      summary: Lihat Semua User (Read)
      tags: [Admin Panel]
      responses:
        200: {description: List user}

  /api/admin/users/{id}:
    delete:
      summary: Hapus User (Delete)
      tags: [Admin Panel]
      parameters: [{in: path, name: id, schema: {type: integer}, required: true}]
      responses: {200: {description: Terhapus}}

  /api/admin/users/{id}/reset-key:
    put:
      summary: Update API Key User (Update)
      description: Endpoint ini menggunakan method PUT untuk memenuhi syarat CRUD Update.
      tags: [Admin Panel]
      parameters: [{in: path, name: id, schema: {type: integer}, required: true}]
      responses:
        200: {description: Key berhasil diupdate}